[update] Rework block moving
This commit reworks the moving of blocks so that they move in relation
to another block rather than a position shift. This addresses some
issues related to the way blocks are stored when shifting across the
children of a parent block.

It also updates the UI such that menu items are disabled rather than
completely removed when inactive.