Allow MenuItem label to be a function (#139)
Problem
I'd like to pass a function to the label prop of MenuItem that could dynamically generate the button text. Currently, the menu item button text can only be set statically and doesn't get re-rendered when the properties of my component block changes. It's inconsistent with other props like `isDisabled` and `onClick`

Solution
In MenuItem, check if the label is a function, and if so call it and return the result as the text in the button. Pass the CK instance, Block, and MenuItem to the callback as arguments just in case they're needed.