[WIP] Add `blockTypesData` for additional props.
Specify (optional) blockType-specific data to pass
as props on render.

# Motivation

Imagine a component that selects from a list:

```
{
  id: 'animals',
  label: 'Animals',
  component: Animals
}
```

You use 'animals' in many places. But then a new 
component request comes in: it should be similar 
to  'animals', but filter results to show only 
warm-blooded animals. 

Now you could create a brand new
blockType, and that might be appropriate in 
certain situations. But in many cases, it would be 
simpler to reuse component logic with props.

# Proposal

Use a new bootstrap options key, `blockTypesData`,
as a map of `blockType.id` keys. The value for
each key is an object, which becomes a part of the
`BlockType` as `data` and spread as props for the
component.

I'm open to approaching this in other ways. This 
is a first pass that I hope helps to communicate 
the end goal.